Update Azure login to use secrets directly (#6)
This pull request makes a minor update to the Azure login step in the
`terraform-plan.yml` GitHub Actions workflow. The change simplifies how
Azure credentials are referenced by using the `${{ secrets.* }}` syntax
directly in the `with` block, removing the unnecessary use of
environment variables.

- Simplified Azure credential usage in the `Azure Login (OIDC)` step by
referencing secrets directly in the `with` block, eliminating redundant
environment variable assignments.