Added environment variable to variables file (#79)
This pull request adds a new variable to the Terraform configuration for
improved environment management. The most important change is the
introduction of an `environment` variable, which allows specifying the
deployment environment (such as dev, staging, or prod).

Environment configuration:

* Added a new `environment` variable to `variables.tf` with a default
value of `"dev"` and a description to clarify its purpose.